requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
psutil>=5.9.0 